Add alpha test for object animations;

+++ DHConcurrentAnimation/Animations/ObjectAnimation/ClothLine/ObjectClothLineVertex.glsl
@@ -5,7 +5,7 @@ uniform vec2 u_offset;
 uniform float u_percent;
 uniform mat4 u_rotationMatrix;
 uniform float u_transitionRatio;
-uniform int u_event;
+uniform float u_event;
 
 layout(location = 0) in vec4 a_position;
 layout(location = 1) in vec2 a_texCoords;
@@ -15,7 +15,7 @@ out vec2 v_texCoords;
 vec4 updatedPosition() {
     vec4 position = a_position;
     position = u_rotationMatrix * position;
-    if (u_event == 0) {
+    if (u_event == 0.f) {
         if (u_percent < u_transitionRatio) {
             position.xy += u_offset * (1.f - (u_percent / u_transitionRatio));
         }
